+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `showOnMobile` on `SearchResult`
+
 ## [1.5.4] - 2019-08-06
 
 ### Changed